updated README, moved hackathon list officers to data
